+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 **Added**
 - Activated GitHub code scanning alerts
+- New option `hide_notebook_metadata` to encapsulate the notebook metadata in an HTML comment (#527)
 
 **Changed**
 - Install Jupytext from source on MyBinder to avoid cache issues (#567)